Release of 2.0.3
Officially 2.0.3 is released, here are the following changes:
Save system was rearranged to prevent accidental button presses of the delete button and better differentiate the sheets and saves. Ideally this should be the last time the save system is moved around.
More checks are in place for the sheets and maths to prevent stupid or otherwise buggy names that can break the code, with an associated error message.
Added modulo function, refer to documentation to determine what it does.